Restore connection broadcast messages.
At some point, I noticed that 'connect' on the client was being
triggered by the actual client's connections, not the broadcast
message I'd specified from the server.

This change revives the old 'new user' event as a broadcast,
allowing all clients to get application data about the user,
instead of a bare connect.